requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
passlib[bcrypt]
paho-mqtt
celery
redis
reportlab  # or weasyprint
